Backend/requirement-amd64-cnd.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
bcrypt=4.0.1=pypi_0
brotlipy=0.7.0=py39h9ed2024_1003
ca-certificates=2023.01.10=hecd8cb5_0
cachetools=5.3.1=pypi_0
capstone=5.0.0rc2=pypi_0
certifi=2023.5.7=py39hecd8cb5_0
cffi=1.15.1=py39h6c40b1e_3
charset-normalizer=3.1.0=pypi_0
click=8.0.4=py39hecd8cb5_0
colored-traceback=0.3.0=pypi_0
cryptography=39.0.1=py39hf6deb26_0
exceptiongroup=1.1.1=pypi_0
fhirclient=4.1.0=pypi_0
fhirpy=1.3.1=pypi_0
flask=2.2.2=py39hecd8cb5_0
flask-cors=3.0.10=pypi_0
flatbuffers=1.12=pypi_0
frozenlist=1.3.3=pypi_0
gast=0.4.0=pypi_0
google-auth=2.19.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.2=pypi_0
h5py=3.8.0=pypi_0
idna=3.4=py39hecd8cb5_0
importlib-metadata=6.6.0=pypi_0
iniconfig=2.0.0=pypi_0
intervaltree=3.1.0=pypi_0
isodate=0.6.1=pypi_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py39hecd8cb5_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
libclang=16.0.0=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
mako=1.2.4=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.3=pypi_0
multidict=6.0.4=pypi_0
munch=3.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
ndjson=0.3.1=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=1.1.1t=hca72f7f_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.2=pypi_0
paramiko=3.2.0=pypi_0
pip=23.0.1=py39hecd8cb5_0
pluggy=1.0.0=pypi_0
plumbum=1.8.2=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.5=pypi_0
pwntools=4.10.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyelftools=0.29=pypi_0
pygments=2.15.1=pypi_0
pynacl=1.5.0=pypi_0
pyopenssl=23.0.0=py39hecd8cb5_0
pyserial=3.5=pypi_0
pysocks=1.7.1=py39hecd8cb5_0
pytest=7.3.1=pypi_0
python=3.9.16=h218abb5_2
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=hca72f7f_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
ropgadget=7.3=pypi_0
rpyc=5.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=67.8.0=py39hecd8cb5_0
six=1.16.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.41.2=h6c40b1e_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow=2.9.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
tomli=2.0.1=pypi_0
tqdm=4.65.0=pypi_0
typing-extensions=4.6.3=pypi_0
tzdata=2023.3=pypi_0
unicorn=2.0.1.post1=pypi_0
urllib3=1.26.16=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py39hecd8cb5_0
wrapt=1.15.0=pypi_0
xz=5.4.2=h6c40b1e_0
yarl=1.9.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h4dc903c_0
